#1990 move selection and deslection to side effects
